added review adjustments
